[IMP] base: add the correponding value field in tree view
In order to display the value of ir_properties in the tree view, we have
to check that all of them have the same type and only then can we add
the right value field in the view.

This is done using a xpath expresssion and only added if all the
ir_properties records that we have are of the same type.