[IMP] mail: do not update the existing subtype on change of parent
purpose of this commit is, when user changes existing record's parent then it
should not update subtypes according to the new parent.

so, only update subtypes at the time of creation otherwise skip.

after this commit, when the user changes a parent of any existing record
then he/she will be not subscribed according to the new parent.

task-1974055