[IMP] web: show start time of regular event in calendar
Purpose of this commit is, to show start time in month and week mode
for regular events(not multi-days), also if hide_time is defined then
do not show time in month mode on event.

After this commit:
For normal event(not multi-days), starting time of an event will be
shown in month mode in calendar, if hide_time is explivitly defined on
calendar view then time will not be displayed on event in month mode,
hide_time now hides time in calendar popover as well as hides time on
calendar event in month mode.

task-2058477

Co-authored-by: Mohammed Shekha <msh@odoo.com>