[FIX] mail: faster opening of messaging menu
Before this commit, opening the messaging menu might be slow.

This was caused by the parsing of messages' bodies with lots of HTML
nodes. Indeed, messages' bodies are linkified and inlined, and both
processes were very slow in their implementation: each node in the
HTML were recursively parsed then resetting 'innerHTML'. Setting the
property 'innerHTML' on a Node triggers a parse HTML, which is quite
slow on complex nodes.

This commit fixes the performance issue:

- with 'inline': by using textContent instead of
  parsing the html content
- with 'linkify': by mutating node only when linkifying a text node.

Task-Id 2076299