[FIX] sale_product_configurator, website_sale(_wishlist): fix multicompany issue
Currently, we are allowing to select the other company's product as
an alternative product or accessory products on the product.

For example,
	website: W1, W2
	company: C1, C2

	W1-> C1
	W2-> C2

	Product: P1
	===========
		company: C1

	Product: P2
	===========
		company: C2
		alternative product: P1

And this will give a traceback because we are going to use another company's
products. And this is obviously a wrong behavior.

That's why put a specific domain on the alternative, accessory and optional
product fields to fix this multi-company issues.

Now product wishlist will be also a company-specific.

task-2070522