[FIX] website_slides: Fix reordering of slide after d&d
This commit fixes the reordering of slides in case there are some
uncategorized slides in the course.

This issue was due to the fact that in a previous commit:
  c4eba6f5b150a0e881b7fb9b935ef1a0c3154bbf

uncategorized slides were moved to the bottom of the list,
which broke the reordering as we use the dom order itself
to reorder the slides with a simple rpc call to dataset/resequence.

Furthermore, having the uncategorized slides at the top of the list,
allows us to be coherent with the backend behaviour.

To avoid future problems, a test was added for _get_categorized_slides

LINKS:

TaskID: 2090927
PR: #39086