[FIX] website_slides: Fix add slide in frontend
If you create two empty categories then try to add in the first empty one,
the slide would be added to the second one instead.

This commit fixes that by using the category_id received from the frontend
instead of relying on the one on the slide itself.

Category_id will no longer be set in the controller
before the resequencing occurs. In fact, category_id will be set
no matter what during the resequencing as it will update the slide
sequence.

LINKS:

TaskID: 2090927
PR: #39086