[FIX] mail: reset suggested partner when record saved
After you once click on 'Send Message' of composer, if you change
the details of the partner(e.g email address), you will have to reload
the page in order to actually get the modified details of follower in
the chatter

We fixed this so that there won't be a need to reload the page for
getting the modified data.

After this commit, when you click on 'Send Message' and then modify
details of partner and save it, it will reset the suggested recipients.
So now when you again click on 'Send Message', it will call
get_suggested_recipients method again to get modified data.

task-2089816

Co-authored-by: Mohammed Shekha <msh@odoo.com>
Co-authored-by: Alexandre Kühn <aku@odoo.com>