[IMP] portal: sending messages without reloading page
before this commit: when the user has sent the messages from the portal chatter,
the page was reloading.

in this commit, improved portal chatter for messaging and attachments.
when the user sends a message, the new message is updated in history without
reloading the page instead of reloading the page as previously and update
test case regarding the changes

task-2054662

Co-authored-by: jpr-odoo <jpr@openerp.com>