[REF] sale_(project, timesheet): generate task from SOL without timesheet
Purpose of the task is, it is not possible to generate a task from an
Sales order line if timehseet is not installed.
However, the user may want to generate a task without necessarily
timesheeting on it (e.g. a task in Field Service).

so split the sale_timesheet module by another module as sale_project
which only depends on sale_management and project.

Task-2123707
Closes: #40594