[REF] survey, hr_recruitment_survey: on survey.survey, remove unused category field
PURPOSE

As new features are about to land in survey, notably live interactions [1]
and new survey building [2] performing a pre cleaning is necessary.

Purpose is to perform some cleaning in survey models: removing unnecessary
fields, performing some cleaning and finally renaming models. Indeed survey
models are known to be quite badly named and difficult to understand.

SPECIFICATIONS

Survey model holds a ``category`` field whose purpose is to be able to somehow
categorize surveys according to their use. However using this field is not
easy as it is hidden and is a simple selection field. Module should add their
own key. Its sole use is in ``hr_recruitment_survey`` which is a niche module.
Let us clean models and lessen model complexity.

LINKS

[0] Related to Task ID 2061901 (survey models cleaning and preparation)
[1] Task ID 1972640 (live interactions)
[2] Task ID 2119587 (new frontend for building surveys)

PR #40765