[REF] survey: on survey.survey, rename certificate field to certification
PURPOSE

As new features are about to land in survey, notably live interactions [1]
and new survey building [2] performing a pre cleaning is necessary.

Purpose is to perform some cleaning in survey models: removing unnecessary
fields, performing some cleaning and finally renaming models. Indeed survey
models are known to be quite badly named and difficult to understand.

SPECIFICATIONS

All certification related fields on survey model begin with certification_ .
Only the boolean one telling if a survey is a certification or not is called
certificate. In order to ease grep and ordering it is renamed to certification.

LINKS

[0] Related to Task ID 2061901 (survey models cleaning and preparation)
[1] Task ID 1972640 (live interactions)
[2] Task ID 2119587 (new frontend for building surveys)

PR #40765