[IMP] crm_livechat: create a lead when operator not giving reply
Currently, a visitor is not notified if the operator does not give a reply.

After this commit:
when an operator won't give reply to a visitor within 30 mins, then a visitor
will be notified by asking their name and email.  after that lead will be
created for that particular visitor and then a visitor will get a confirmation
mail that he/she will soon get contacted.

task-2030381