[REF] stock: perf orderpoint as batch
Currently orderpoint run one request at a time. However it's not
efficient since compute are some private methods could be executed
one time in batch instead of record by record.

Instead of a call to run each time a procurement request is ready,
append it to a list and call run with the list of procurements.

Task: 1849912