[IMP] mrp: add extra information about scheduling
Add scheduling information (icon) with popover to show info/warning/error
about the scheduling of workorder. Four extra possible pieces of information
is given (level change the color of the icon):
- Info level; If previous WO and pending state, show sheduling info of the previous WO
- Warning level; info if the WO is late
- Error level; if conflit with other WO (same workcenter in same time)
- Error level; if previous WO is planned after the current one

Also, add a button (in the popover) to replan correctly all WO of the MO of
the current WO (only if it is not in progress and errors is detected).

All is done via a JSON data compute field ('json_popover') in the workorder.
It is allowing use to make computation in batch and avoid to add lot of fields.

task-2082261