[ADD] web: add user-times icon into svg and png formats
TaskID: 2060206

When a record is unassigned, we display the user-slash by default.
But if we want to use this icon instead of the user-slash icon.
We just change the name of file into the _get_placeholder_filename
in res.users model.
(file : odoo/odoo/addons/base/models/res_users.py)