[IMP] survey:  Illustrate your questions and answers with media
Now you can add media to a question in order to illustrate it. This option is available in the
tab "Options" when you create a question under the section "Illustration".
You can add multiple media and organize them as you want. They will be displayed in a carousel
on the frontend, below the question and a potential description.

You have also the possibility to add pictures on an answer for mcq if it makes sens with the
display mode, this option is available in the tab "Options" under the section "Display Mode".

Demo data has been add to test these new features.

task-2083823