[IMP] mail: add feature to modify the bounced mail content for alias
In order to improve the flexibility of alias, add a
customisable html field on the alias model.
This html content will be send (in a email) in case of
bounced/unauthorized mail recieved for this alias,
it is not available when Alias Contact Security is on
'everyone' because not access error is possible in this case.
If it is not set a default generic mail will be send
depending of the contact security.
Also, add a references message-id of the mail received to the answer.

task-2126509