[FIX] stock: show move lines with quantity done
Before this commit, pickings of picking type with `show_reserved` on
False don't show pre-fill moves -what is expected- even if those
pre-fill moves have quantity done -what is less wanted and can happened,
with stock_barcode for example.

With this commit, we continue to hide pre-fill moves for this kind of
picking, except if for pre-fill moves who have a quantity done.