[FIX] hr_expense : fix issues in hr_expense in a multicompany situation.
The company_id wasn't propagating correctly from the expense or expense report
to the invoice and payment.

The employee_id was prefilled automatically with the one linked to the user. This caused issue
when a manager wanted to create an expense for an employee in a different company.

This commit fixes those issues and also introduces a unit test to ensure that it keeps working.

Task-2088891