[FIX] project : fix multicompany issues with projects.
If a task and its parent task were in different companies, there were issues with the
sub-tasks and parent tasks smart buttons.

It was possible to create an attachment with a different company_id than the one from the
project even if the user wasn't showing the records from this company.

This commits fixes this and introduces a test.

Task-2088891