[FIX] web: boolean widgets in list views
Before this commit, pointer events were prevented on all checkboxes in readonly
list views. That meant that the boolean_toggle widget for instance could not be
properly toggled in a non-editable list (which was the purpose of such a widget).

Now, pointer events are only disabled if the checkbox is not a widget field,
meaning that any widget applied to a boolean will delegate the handling of all
events to its widget.

This fix cannot be tested as the `pointer-events` property does not affect manual
event triggering.

Task 2154055