[FIX] web_editor, website_forum: image drag and drop into forum's text editor
Since 62c9dedafda0cbdc618641eaa80c49535ab0b5f6, if a portal user drag
and drop a picture into the forum's text editor, the image is never
loaded and an error is shown in the logs (right access error).

Now, the portal user is allowed to create attachments for the forum.

opw-2127345