[IMP] website_blog: move latest posts snippet option in standard
Previously, the "list-style" and "first is bigger" snippet options for
the latest_posts_big_pictures snippet were only available in the
graphene and avantgarde theme, this commit moves these options into
website_blog so that they are available in every theme. It also fixes
the list-style animations as they were all broken since the snippet's
XML was changed. Since the DOM of the snippet has changed quite a bit
and the size and aspect ratio have been altered, this is a best
rendition of the previous effects and not a perfect 1-to-1 reproduction.