[IMP] portal,portal_rating: make required message more obvious
Current implementation of module 'portal' requires a message in order
for a post to occur (in portal.mail.PortalChatter.portal_chatter_post(...)).
Because module 'portal_rating' inherits this model, when ratings are
submitted in website_slide and website_sale without a message, the
website is refreshed but nothing is saved in the database. This page refresh
makes users think that their rating is submitted even though it is not.

Making the corresponding textarea 'required' for now at least makes this
non-posting behavior more obvious, but is most likely not a long term solution.
Refactoring is needed for a better implementation (e.g. either making
the message not required at the portal method level, separating out rating
logic from portal, or some other similar logic.)